+++ OhHyungDong/_01_LoginForm/src/main/resources/mappers/board/member/memberregist-mapper.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.sh.choichanguk.member.model.dao.MemberRegistMapper">
+    <select id="findByEmail" resultType="String">
+        select member_email
+        from member
+        where member_email=#{email}
+    </select>
+
+    <insert id="insertMember">
+        INSERT INTO member (member_email, member_password, member_name, registration_date,member_profile_url)
+        VALUES
+        (#{memberEmail}, #{memberPassword}, #{memberName}, NOW(),#{url})
+    </insert>
+
+</mapper>
