--- YangHeeyun/board/src/main/java/com/sh/login/BoardApplication.java
@@ -1,13 +1,0 @@
-package com.sh.login;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BoardApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(BoardApplication.class, args);
-    }
-
-}
