--- YangHeeyun/board/src/main/java/com/sh/login/model/dao/MemberMapper.java
@@ -1,13 +1,0 @@
-package com.sh.login.model.dao;
-
-import com.sh.login.model.dto.MemberDto;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface MemberMapper {
-    // 회원가입
-    int insertMember(MemberDto memberDto);
-
-    // 이메일 중복체크
-    int checkDuplicateEmail(String memberEmail);
-}
