--- YangHeeyun/board/src/main/java/com/sh/login/model/service/MemberQueryService.java
@@ -1,18 +1,0 @@
-package com.sh.login.model.service;
-
-import com.sh.login.model.dao.MemberMapper;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class MemberQueryService {
-    private final MemberMapper memberMapper;
-
-    // 이메일 중복확인
-    public int checkDuplicatEmail(String memberEmail){
-        int emailCheck = memberMapper.checkDuplicateEmail(memberEmail);
-        return emailCheck;
-    }
-
-}
