+++ YangHeeyun/sod/src/main/java/com/sh/SodApplication.java
@@ -1,0 +1,13 @@
+package com.sh;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SodApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SodApplication.class, args);
+    }
+
+}
