+++ YangHeeyun/sod/src/main/java/com/sh/board/model/dto/CategoryDto.java
@@ -1,0 +1,6 @@
+package com.sh.board.model.dto;
+
+public class CategoryDto {
+    private int categoryCode;
+    private String categoryName;
+}
