+++ YangHeeyun/sod/src/main/java/com/sh/member/model/dao/MemberMapper.java
@@ -1,0 +1,13 @@
+package com.sh.member.model.dao;
+
+import com.sh.member.model.dto.MemberDto;
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface MemberMapper {
+    // 회원가입
+    int insertMember(MemberDto memberDto);
+
+    // 이메일 중복체크
+    int checkDuplicateEmail(String memberEmail);
+}
