+++ YangHeeyun/sod/src/main/resources/mappers/board/board-mapper.xml
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.sh.board.model.dao.BoardMapper">
+    <!-- 전체 게시물 조회 -->
+    <select id="findAll">
+        select
+            *
+        from
+            board b
+            join category c using(category_code)
+            join member m using(member_id)
+
+        where
+            board_status = 'Y'
+        limit
+            #{offset}, #{limit}
+    </select>
+    <select id="count" resultType="_int">
+        select
+            count(*)
+        from
+            board
+        where
+            board_status = 'Y'
+    </select>
+</mapper>
