+++ YangHeeyun/sod/src/main/resources/mappers/member/member-mapper.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.sh.member.model.dao.MemberMapper">
+    <!-- 회원가입 -->
+    <insert id="insertMember" useGeneratedKeys="true" keyProperty="memberId">
+        insert into
+        member(member_email, member_password, member_name, profile_path)
+        values (
+        #{memberEmail}, #{memberPassword},#{memberName}, #{profilePath}
+        )
+    </insert>
+    <!-- 이메일 중복체크 -->
+    <select id="checkDuplicateEmail" parameterType="string" resultType="int">
+        select
+        count(*)
+        from
+        member
+        where
+        member_email = #{memberEmail}
+    </select>
+</mapper>
