+++ YangHeeyun/sod/src/main/resources/templates/board/boardList.html
@@ -1,0 +1,116 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>게시판</title>
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+    <style>
+        body {
+            display: flex;
+            flex-direction: column;
+            min-height: 100vh;
+        }
+        main {
+            flex: 1;
+            padding: 20px;
+        }
+
+        .pagination-container {
+            display: flex;
+            justify-content: center;
+        }
+        tr[data-href] {
+            cursor: pointer;
+        }
+    </style>
+</head>
+<body>
+<nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+    <a class="navbar-brand" href="#">Spring or Death</a>
+    <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
+        <span class="navbar-toggler-icon"></span>
+    </button>
+    <div class="collapse navbar-collapse" id="navbarNav">
+        <ul class="navbar-nav mr-auto">
+            <li class="nav-item">
+                <a class="nav-link" href="#">홈</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="#">게시판</a>
+            </li>
+            <li class="nav-item">
+                <a class="nav-link" href="#">기타</a>
+            </li>
+        </ul>
+    </div>
+</nav>
+<main class="container">
+    <div class="search-container">
+        <form class="form-inline">
+            <input class="form-control mr-sm-2" type="search" placeholder="검색어를 입력하세요" aria-label="Search" id="searchInput">
+            <button class="btn btn-outline-success my-2 my-sm-0" type="button" id="searchButton">검색</button>
+        </form>
+    </div>
+    <div class="table-container my-3">
+        <table class="table table-hover">
+            <thead class="thead-light">
+            <tr>
+                <th>글번호</th>
+                <th>카테고리</th>
+                <th>글제목</th>
+                <th>작성자</th>
+                <th>조회수</th>
+                <th>작성일</th>
+            </tr>
+            </thead>
+            <tbody id="tableBody">
+            <tr th:each="board, stat : ${boards}" th:object="${boards}">
+                <td th:text="${board.boardNo}"></td>
+                <td th:text="${board.categoryCode.categoryName}"></td>
+                <td th:text="${board.boardTitle}"></td>
+                <td th:text="${board.member.memberName}"></td>
+                <td th:text="${board.boardCount}"></td>
+                <td th:text="${board.createdAt}"></td>
+            </tr>
+            </tbody>
+        </table>
+    </div>
+    <div class="pagination-container">
+        <nav>
+            <ul class="pagination">
+                <li class="page-item disabled">
+                    <a class="page-link" href="#" tabindex="-1">이전</a>
+                </li>
+                <li class="page-item active">
+                    <a class="page-link" href="#">1 <span class="sr-only">(current)</span></a>
+                </li>
+                <li class="page-item">
+                    <a class="page-link" href="#">2</a>
+                </li>
+                <li class="page-item">
+                    <a class="page-link" href="#">3</a>
+                </li>
+                <li class="page-item">
+                    <a class="page-link" href="#">다음</a>
+                </li>
+            </ul>
+        </nav>
+    </div>
+</main>
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.16.0/umd/popper.min.js"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
+<script>
+
+    document.addEventListener('DOMContentLoaded', function() {
+        const rows = document.querySelectorAll('tr[data-href]');
+        rows.forEach(function(row) {
+            row.addEventListener('click', function() {
+                window.location.href = row.getAttribute('data-href');
+            });
+        });
+    });
+</script>
+</body>
+</html>
